Fix for subcircuit multipliers for MULT parameter
This is related to this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/889

This was reported by an external user.

The PSP device model uses MULT instead of M for its multiplier
parameter.    For Xyce, we hack in an alias for M into the PSP, but not
everyone will do this (for example when dynamically linking in a new
ADMS-generated version of the PSP).

The subcircuit multiplier code was written in Xyce to only apply
subcircuit multipliers named "M".  So, it overlooks aliases like MULT.
The subcircuit M parameters are handled as a special case since they are
multplicative thru the subcircuit hierarchy.

This commit modifies the subckt M code to also treat MULT as a
multiplier (but only for device model params).