Refactor and fix iterative refinement solver
The iterative refinement solver in Xyce was written to reuse previous LU
factors in subsequent linear solves to expedite linear solver time.  It is
very rudimentary and, as it turns out, incorrect in circumstances where
transformations are being used on the linear system (like singleton removal).
That has been fixed, while the code has been refactored for ease of understanding.

The solver has also been changed to take into account the current Newton iteration
in determining whether to refactor the matrix or attempt to use iterative refinement.
This will be the first solver to include the nonlinear solver information to make
linear solver decisions.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/894