Adds option for changing minimum tolerance for iterative refinement
The iterative refinement solver will only try reusing old solver information
if the old solver can reduce the relative residual of a new linear system by
something smaller than the minimum tolerance.  This can be adapted for the
circuit of interest, so it should be exposed in the linear solver options.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/894