Partial implementation of new .SENS param spec
This is in support of an LDRD project, but it may also be generally
useful for other use cases.

Historically for .SENS, the way Xyce works is that it requires the  user
to specify a list of parameters, which are the subject of the
sensitivity analysis.   This is on the .SENS line as
"param = p1,p2,p3,..."

This is fine for some cases, but it can be burdensome when dealing with
really large numbers of parameters.  Also, in some cases (as is the case
for the LDRD project) the parameters of interest are not traditional
device parameters, nor are they from .param statements.  The .SENS
analysis relies a lot on the parameter handling structure of the device
package in Xyce.  This is very convenient but unfortunately it means
that for parameters that DON'T use this machinery, they are excluded
from sensitivities.

In the case of the LDRD, they come from an external (non-netlist)
file.  So, I needed to write special code for them.  That is the subject
of this commit.   I think, however, that the utility of this code will
go well beyond the LDRD.

This is a checkpoint commit.  This work isn't done yet.