Use circuit input file directory in Simulink initialization
When Simulink is running Xyce the current directory in Matlab
may not be where Simulink was set to find the circuit input
file. This would cause Xyce to exit early because it could
not find the input circuit file.  Now the Simulink layer
funtions XyceSimMask.m and XyceSFunctionImp.m use the
directory where the circuit input file was selected as
a working directory for Xyce simulation without chaning
the Matlab working directory.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/883