Implement device-wide .SENS param spec
This is in support of these two issues:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/910

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/901

Xyce sensitivity analys historically requires the user to explicitly
specific all the sensitivity parameters in the netlist.  For large
parameter sets, this can be a headache.   Also, there exist device
models which specify parameters in a non-traditional way, and thus don't
naturally plug into the existing parameter data structures in Xyce.

So, in some cases a different specification is more convenient, and in
other cases it is the only possible way to set up the analysis.

This commit implements a new waay of specificying sensitivity
parameters.  Historically for Xyce .SENS, parameters are specified
on the .SENS line as "param = p1,p2,p3,..."

Now, there is an alternate method, where the user can instead say (again
on the .SENS line);   SENSDEVICENAME=name, where "name" is the name of
the device.  For instrumented devices they can then provide a full list
of their parameters for the analysis.

This is set up now, and works, at least for direct sensitivities in DC
and transient.  It has not yet been set up for AC sensitivities, which
are implemented in a different part of the source code.  It also doesn't
work in parallel yet, but I think that won't take long to fix.