Fix unit test issues for parallel builds
Some unit tests are not set up to work in parallel
(i.e. XyceCInterface and the parserUnitTest).  These
tests are still run in serial.  The issue with
XyceCInterface is that the uderlying C-Interface isn't
designed to take an MPI pointer so it can't start Xyce
in parallel.  parserUnitTest is more complicated but
has a dependence on newExpression being compiled
to take double type that results in duplicate symbols
when linked to XyceLib.  Both of these tests are
still run in serial.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/909