Fix adjoint sensitivity for "device level" spec
This is in support of these two issues:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/910
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/901

For my previous work on the "device level" parameter specification,
I only focused on transient direct sensitivities.  This made sense as
they are easier to set up, and 99% of the work necessary for them would
be applicable to adjoints.

However, that wasn't quite enough for adjoints to work correctly.  This
is mostly because they rely on a different set of functions and they
needed to be modified to call the optional new way of setting up the
"function derivative"  (i.e. the dfdp terms coming from the device
model(s)).

This commit is a partial fix.  I've created three regression tests for
this capability, and at the moment only one of them passes.