Final fix for adjoints with device-level spec
This is in support of these two issues:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/910
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/901

This final fix is for a specific use case, when
(1) the device has time-dependent parameters
(2) the time dependent parameters matter to the Jacobian

When doing transient adjoints, the solutions are saved, but the
Jacobians are not, from the forward calculation.  So, the Jacobians are
re-computed on the fly when doing reverse time integration for adjoints.

In the use case above, the time dependent params were not being updated
properly during the reverse integration, leading to incorrect Jacobians,
and thus incorrect adjoint results.   Fixed now.