Fix device-level sensitivity spec for parallel
This is in support of these two issues:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/910
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/901

Up until this commit, my work on the new method for specifying
sensitivity parameters was only tested in serial.  I was pretty sure
that to make it work in parallel, I only needed to fix a single
function, and that proved to be correct.  This commit fixes that single
function.

The specific single function is a new one that was added to the device
package to support this feature, DeviceMgr::getSensParamsForDevice.
That function finds the specified device, and then returns vectors of
names and original values for those parameters.  In parallel, that
device only exist on a single processor, so those vectors have to be
broadcast to the other processors.  Currently, in the sensitivity code,
it is necessary for every processor to know the full list of parameters.