add notify states; add them to diagnostic output
An external user today pointed out that the diagnostic output, specified
via .options DIAGNOSITIC never confirms if DCOP calculations succeeed or
not.  This capability only reports if the phases of DCOP have failed.

This was relevant b/c the user was running an AC calculation with .STEP,
and the results were not changing from one iteration to the next.  The
outer loop iteration was over a Vsrc value.  It seemed possible that the
lack of change was due to the DCOP calculation not being re-performed
(or possibly failing).  The user attempted to confirm this via
DIAGNOSTIC, and it wasn't clear.

This commit adds more DIAGNOSTIC outputs, corresponding to SUCCESS of
the various DCOP phases (std Newton, gmin stepping and source stepping).