fix uninitialized return variables.
The getMutuableValue() methods were declaring a return
variable but not setting an initial value.  If the conditional
logic of the routine didn't set the value then an uninitialized
value would be returned causing the code to error out for
some newer compilers. The fix was simple -- just initialize
return variables when they are declared.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/953