Add release note for deprecating autotools
There is development work in progress that will break backwards
compatibility with older versions of Trilinos. Once that is brought into
the main branch, it will not be practical to continue supporting the
AutoTools build system for Xyce. The release note update is intended
to provide notice of this feature change.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/902