Make MPI bug test a unit test
This small MPI bug test runs a small program to check
if the MPI implementation has a particular bug.  This was
called from the regression suite by guessing from the
binary path to locate the test program.  It's much
easier to just make this a unit test and avoid the guessing.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/877