Draft updates to CMake install instructions
CMake is the official build system for Xyce 7.10 and support for autotoo
ls is being deprecated. The CMake infrastructure has been in place (and
used for development) for a while, but the build instructions still
reference autotools as the primary build system. This commit incudes a
first pass at updating the information in the build guide documents.
This first pass mostly focuses on details for Windows systems.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/902