Small improvement to TESTJAC diagnostic
I was using the TESTJAC diagnostic today to track down a problem in a
device model.

While doing it, I noticed that if there was a failure in either the F
matrix *or* the Q matrix, it would output that both the dFdx and dQdx
comparisons had failed.  But frequently, only one or the other failed.

This commit modifies the capability to have two separate booleans rather
than one.  One is for F and the other is for Q, so the output will make
a little more sense.