Update configure to label the release version as (Q)Release
We're working on some procedure changes for this release (including how
the release is labeled), and we intentionally did not set the phase/state
flag on the release branch. However, there are some unexpected
documentation and testing changes taking place in this release. These
other updates are going to require more builds and testing from the
release branch than planned. To help manage the extra builds, this
commit sets the qualification flag in the build system.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/965