Use Xyce_QA as the CTest project on release branch
Regression testing of a release branch reports to a different CDash
dashboard. The legacy build scripts use the branch parameter or
environment variable to indicate which dashboard to report to. The
initial development of the GitLab pipeline test job did not include this
check; the GitLab pipeline jobs always reports to the main development
dashboard. This commit adds the branch check to the ctest process;  the
ctest drop location is set to the ctest project name, and if the
branch name starts with 'release' the ctest project is set to Xyce_QA
(set to Xyce otherwise).

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/968