Clean, unmodified bsim soi v4.7.0 *va file
This is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/958
"Implement the BSIM SOI v4.7.0 in Xyce"

This will get modified to work with ADMS.  But, starting with the
unmodified file so that it is easy to track the changes.

The README_Xyce file is (for now) copied from 4.6.1.  In the text, the
numbers 4.6.1 have been updated to 4.7.0, but otherwise no changes yet.
It is likely that the README_Xyce documentation will be nearly the
same.