Temporarily add constants
This is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/958
"Implement the BSIM SOI v4.7.0 in Xyce"

This commit adds two constants.  These are part of the newest LRM
constants.vams file.  But, I'm using an older version of ADMS, with an
out-of-date constants file, so for now I'm adding them here.   They can
be removed later.

The two added constants are   P_Q_NIST1998   and   P_K_NIST1998.