Fix verilog so it now compiles with ADMS
This is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/958
"Implement the BSIM SOI v4.7.0 in Xyce"

This commit modifieds the verilog-A file such that it will now compile
in ADMS and produce the *C, *h files.

Most of these changes were copied from the "diff" for the version 4.6.1
of this device.  I have not tested this to see if it works yet.

I have also not attempted to add any of the "multiplier" changes.