Complete updates to BSIM SOI 4.7.0 verilog file
This is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/958
"Implement the BSIM SOI v4.7.0 in Xyce"

The main change to the verilog in this commit is to add
multipliers to it.  Most of the other necessary changes were
in my previous commit.  The *va code compiles without much fuss.

Also, the resulting *C and *h file compile fine, and I have added them
to the Xyce source.    The CMC tests for this device mostly pass.
The ones that don't are off by small amounts (which was the case
for previous BSIM SOI versions).

The previous BSIM-SOI version, v4.6.1 has been updated to have a new
level number.  It is now level=70461 rather than level=70.  The v4.7.0
is now level=70.  This follows the previously established pattern, where
version 4.5.0 used level=70450.

In order to move the 4.6.1 to a new level number, it was necessary
to modify the verilog file for that device, and regenerate it.  In
addition to the level number itself changing, the filenames needed to be
changed as well, so as not to get name conflicts with the new device.

So, previously, the 4.6.1 model was in the N_DEV_ADMSbsimsoi.C/h files.

Now, v4.6.1 is in the N_DEV_ADMSbsimsoi461.C/h files, and the v4.7.0 is in
the N_DEV_ADMSbsimsoi.C/h files.