Undo redundant installation of admsXml
Previously, when the ADMS capability was enabled, the CMake would find
the admsXml program and copy it into the same bin directory as Xyce.
This, obviously, could cause problems; and it is completely unnecessary.

This commit reverses that mistake and properly has the CMake-based
plugin compiler find the installed admsXml. It looks for it in the same
place that Xyce found it at configure time.

I also made one other change. Some time back, the XyceConfig.cmake.in
file was changed so that it did not know where to look for Trilinos. I
re-enabled that capability such that people should no longer need to
point to the Trilinos installation when linking to Xyce via a CMake
"find" command.

https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/74