Fix ddt terms in bsimsoi v4.7.0
This is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/958
"Implement the BSIM SOI v4.7.0 in Xyce"

In my work up until this point, I neglected one type of change that is
necessary.  ADMS/Xyce is not smart enough to handle contribution lines
that mix together static and dynamic terms.  So, any lines that do this
inside of a verilog-A file, must be modified to separate them into two
separate lines.

So, a line like this:

I(b) <+  A + B + C + ddt(D);

must be changed to:

I(b) <+  A + B + C;
I(b) <+  ddt(D);

This has been completed now, in this device.  One of the regression test
cases, invert_b4soi, did not run in this device before this change.
Now, it runs successfully.