Modify INSTALL.md for clarity and corrections
I found the collapsed sections confusing, and I figured others might
have the same experience, so I made the collapses more obvious. I also
made formatting changes to make it easier to read in a plain text
editor.

The editorial changes are intended to improve clarity for people who are
not experts in coding or numerical methods. I also made a couple of
corrections. In particular, we do test Trilinos 16.1 a little bit, so I
made that the lowest recommended version.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/993