Allow plugin feature to load more than two plugins
The "-plugin" feature is supposed to take a comma separated list of
plugin shared libraries and load them all.  As noted in GH-148 and in
an older Google Groups posting, it only worked for one or two, not for
three or more.

Multiple solutions to this problem have been posted.  This commit
implements a generic "splitArg" function that will split a string on
any given delimeter and return a vector of the split-out strings.

It then replaces the convoluted "-param" loop with a simple loop over
the vector.

Closes GH-148.