Fix capacitance issue in BSIM SOI v4.6.1
This is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/958
"Implement the BSIM SOI v4.7.0 in Xyce"

While this issue is ostensibly about implementing version 4.7.0 of the
BSIM-SOI, the prime short-term  motivation is to fix a known capacitance
bug in version 4.6.1 of the device.

That issue was the first thing fixed in v4.7.0, relative to 4.6.1,
according to the release notes.

I had originally intended for the 4.7.0 implementation to be our fix for
the capacitance problem.  But, unfortunately I encountered some
complications with v4.7.0, related to inaccurate Jacobian derivatives.
Possibly this isn't a problem with the device model itself, but is just
related to our implementation via ADMS.    But at any rate, it is much
more expeditious to fix the capacitance mistake in 4.6.1 than it is
to track down the Jacobian mistakes in the 4.7.0.

Also, it appears that this capacitance mistake has been fixed in the
4.6.1 implementations in at least two other simulators.

So, based on these arguments, the capacitance mistake is now fixed in
the Xyce variant of version 4.6.1, as of this commit.

Most (all?) of our regression tests are not sensitive to the capacitance
mistake.  So, hopefully I won't have to re-generated any BSIM SOI test
 gold standards.  But if necessary I'll do that soon.