Fix incorrect output of FD ASCII rawfile in sweeps
A user reported that when Xyce is asked to produce ASCII
rawfiles (with the "-a" command line switch) and .print statements
with "format=raw", the number of points is incorrect for every plot
but the last if there is a .step statement in the netlist.

Xyce was overwriting the first character of the correct value with
"0".

The same issue is not present for time-domain output.

That's because the doFinishOutput method of the raw ascii outputter
for frequency domain didn't conditionalize output of the number of
points on whether that number was zero or not, as the binary rawfile
outputter does (and the time domain ascii rawfile outputter does).

This commit corrects the issue.

Closes GH-151