Document the BSIM SOI 4.7.0 model
This is in support of this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/958
"Implement the BSIM SOI v4.7.0 in Xyce"

Adding these parameter tables led me to notice an issue in the original
verilog-A for this device model.  The issue was that in many model
descriptions, there were comments directly in the text (using C-style
comment delimeters) to indicate which bugs were associated with certain
parameters.

These comments got passed thru to the Xyce internal parameter tables.
There, they didn't do any harm (except that they added unnecessary text
to the Xyce -param output).  However, they did do harm when I used the
Xyce -doc capability to generate the latex files for the reference
guide.   Latex didn't know what to do with the C-style comment
delimiters and barfed.

To fix this I removed the comments from our hacked version of the
verilog-A file, regenerated the model, and then ran Xyce -doc again.
Once I did that, it worked fine.