Document  capacitance fix to bsimsoi 4.6.1 model
This is primarily in support of this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/988
"Fix the capacitance issue in the BSIM SOI 4.6.1"

but also (to a lesser extent) this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/958
"Implement the BSIM SOI v4.7.0 in Xyce"

The capacitance fix in the bsim-soi version 4.6.1 is now documented in
the fixed defects table.

The BSIM-SOI section of the reference guide has been rewritten,
especially with respect to this issue.  But, while I was doing that
rewrite, I noticed that I never added an output variables table for
the BSIM-SOI version 4.7.0.  This was an oversight that has been
corrected.

To correct it, I needed to create a Makefile for the 4.7.0 model in the
Xyce/utils/ADMS/examples/BSIM-SOI_4 directory, which mimiced the
makefiles in the other BSIM-SOI subdirectories.  The output variables
table is generated via ADMS, if given the correct commands and *xml file
inputs.