Add a new parameter category
This is related to this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/885
"Implement the empirical mu-table method in the rad-aware 1D PDE(TCAD)
device"

The TCAD devices are unorthodox, at least for a circuit simulator, so
sometimes documenting them in a circuit guide is tricky.  Adding a new
parameter category made the documentation slightly easier.

This also includes a very small change to the user guide.