Implement analytical sensitivities in Vsrc,Isrc
This commit is related to this issue
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/849
"Sensitivities don't work with PWL source"

To support sensitivity analysis, it is necessary for the device
under consideration to provide a residual (the RHS contributions)
derivative with repect to the parameter of interest.  This is then
incorporated into the larger sensitivity analysis formula as one
of the terms.

The best option is (when possible) to have a device model compute
this derivative analytically.  Many devices in Xyce have been
instrumented to do this, either using automatic differentiation, or
via manual derivatives.

For devices that do not have analytical derivatives implemented,
most of them can still provide a derivative using finite differences on
the device level.  However, for the finite difference option to work,
it is necessary for the device parameter of interest to be set up
using Xyce's parameter handling structures.

Unfortunately, some devices are unorthodox, so this doesn't work.

The Vsrc and Isrc devices have a slightly unorthodox structure.
Additionally, the PWL sources are even more unorthodox, at least with
respect to parameter handling.  This is b/c PWL sources allow an
arbitrary number of time-value pairs.  As such, for PWL sources, neither
of the two derivative options ( analytical or device-level FD) worked.

For PWL sources, the list of values is handled thru the "V" parameter,
where an arbitrary number of "V" values are given.  But a user
will sometimes want to specify a specific entry in the vector of "V"
values.  The implied "full" name for these parameters is thus V0, V1,
V2, V3, V4, etc.  But if a user specifes(for example) "V3", the code
will exit with error because it isn't in the hash map of params.  Only
"V" is in the container for that device.

This parsing issue is one reason I never implemented dfdp
derivatives for the PWL sources.  But, I hadn't implemented dfdp
derivatives for other types of independent sources either.

This commit implements analytical derivatives for most independent
source types.  That includes PULSE, EXP, SIN, SFFM and PWL.  PWL sources
were the main focus of this issue, but as long as I was in this
part of the code, I decided to implement  derivatives for most
of the common sources.

To address the parsing issue, I've implemented klunky fix, hopefully
temporarily.  The parameter table now has a bunch of placeholder
parameters, going from v0-v9.  So, analytical sensitivities will
work for params v0-v9.  For sources with more than this, any value after
the tenth entry won't work correctly.  This is entirely b/c of parser
issues, and should ultimately be fixable.  The derivative function
itself (in the N_DEV_SourceData file) will  work for more than 10,
if the parsing issues get fixed properly.

I made an initial attempt at a more comprehensive fix for
PWL the parsing issues, so that
parameter names such as "V" (with an implicit index wildcard) would not
cause a "parameter not found" error code and exit.  There are two
functions that need revising in order to fix the parsing.  One is
DeviceEntity::analyticSensitivityAvailable, and the other is
DeviceMgr::getParamAndReduce.

Both of these functions exit with error if a sensitity param
cannot be found.  The first one, which is ONLY used for
sensitivity analysis, has been fixed with a derived class
function SourceInstance::analyticSensitivityAvailable.

As of this writing (August, 2025), the other function
(DeviceMgr::getParamAndReduce), has not been fixed.  This is partly
because it is a much more general function, and is used for a wide
variety of parameter handling purposes, and is thus more dangerous to
fix quickly.  But ultimately it is fixable as well.  But until
it is fixed, the artificial parameters (V3, V4, etc) will need
to be present in the Vsrc and ISRC devices.