Implement device-wide .SENS spec for vsrc,isrc
This is in support of this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/910
"Adapt sensitivity analysis to allow for convenient specification of device-wide parameters"

Also, it is related to this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/849
"Sensitivities don't work with PWL source"

Sensitivity analysis in Xyce is written with traditional,
non-vector, double precision device parameters in mind.  However,
some devices have parameters that, for one reason or another,
fall outside of that paradigm.  For those devices, it can be a
headache to set up sensitivity analysis, as they don't neatly
plug into the default parameter handling structures.

It occurred to me recently that for devices like this, a different
method for specifying sensitivity parameters would be useful.  So,
I set something up.   In this new method, the user can, instead of this:

.SENS objfunc={V(2)} param=vin:par1,vin:par2,vin:par3,vin:par3

do this:

.SENS objfunc={v(2)} sensdevicename=vin

This instructs Xyce to perform sensitivity analysis on all the
parameters associated with the device named "vin".  And, conveniently,
the user doesn't have to specify each individual parameter name (or even
know what they are).

I initially implemented this "device-wide param" sensitivity
specification to support a proprietary model in Xyce.
But I recently remembered that other devices could benefit.

The PWL independent sources are a device where this kind of
specification is convenient.  The reason for this is that PWL
devices have an arbitrary number of parameters, as the list of
entries in the PWL table has arbitrary length.  The parser kind of
treats this as a vector, giving the whole list the same parameter
name.  So, this type of device doesn't easily fit into the
above pattern.

This commit sets up the above capability for Vsrc and Isrc devices.