Switch include<cmath> to include<N_UTL_Math.h>
This is in support of this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/1000
"M_PI definition missing (for an externally reported build)"

For an external user, Xyce didn't build due to this issue.

One of the suggested changes was to replace the include of <cmath> with
<N_UTL_Math.h> in these two files.  They are the only files that
included cmath directly rather than getting it from N_UTL_Math.h.
For M_PI, N_UTL_Math.h has a fallback definition for M_PI.