Fix -param, to return info for specified device
This is a fix for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/990
"Usage text misrepresents how the "-param" feature can be used"

An external user pointed out that when "Xyce -h" is run, the behavior it
describes for Xyce -param is not correct.   It claims that the user can
optionally have device and level arguments after the "-param" argument,
and then Xyce will print params for ONLY that device.

In other words, based on Xyce -h, this should work:

Xyce -param M 10

and Xyce will only print out the level=10 MOSFET device parameters,
rather than the full set of parameters for all devices.

Instead, if the user specified any arguments after "-param", they were
ignored.

The reason for this behavior dates back to a bugfix in 2006.  A user
wanted Xyce -param to return a parameter list, even if the user
specified a netlist on the line.  There is no easy way for Xyce
to differentiate between a netlist name and a device name.  So, only one
of the two use cases could be made to work, and we chose to honor
that user request.

However, reading over the old bug associated with this (1132), it is
clear that the user really wanted what is now supported by Xyce -syntax,
and/or Xyce -norun.  ie, they wanted diagnostic information about
the specified circuit, including device count.  So, I think that
fixing Xyce -param in this way was a mistake.

This commit restores the behavior described by Xyce -h for Xyce -param.