Create new "replaceDependentParameter" function
This is in support of this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/1006

"Fix runaway memory problem for rad function types other than 7"

Many device models internally create transient source functions by
converting them internally into expressions, which are then attached to
device parameters.   These were then added to the parameter handling
capability of Xyce by calling function DeviceEntity::setDependentParameter.

Unfortunately, this function didn't check if the passed parameter
already existed or not, and unconditionally added it to a std::vector
object.  This wasn't a problem if "setDependentParameter" was only called
once for that parameter, but there are use cases in which it is called
multiple times.

I didn't necessarily want to add extra expense to setDependentParameter,
so to fix this I've created a new function
"DeviceEntity::replaceDependentParameter".  This function is now called
from the devices that need to use it (which are proprietary device
models, mostly).