Update DOI and DOECode links
The DOI in README.md points to a very old DOECode page that has an
incorrect link to a Github repository that isn't the Xyce team's and
which contains an ancient version of Xyce 6.2 that hasn't been touched
in 11 years.

This commit updates the DOE Code link and the DOI to point to a much
more recent version.

Closes GH-169